+++ lib/infopark_kickstarter/rake/integration_task.rb
@@ -56,6 +56,7 @@ module InfoparkKickstarter
           sh "cd #{app_path} && bundle exec rails generate cms:component:language_switch"
           sh "cd #{app_path} && bundle exec rails generate cms:component:profile_page"
           sh "cd #{app_path} && bundle exec rails generate cms:widget:google_maps --cms_path=/website/en/_boxes"
+          sh "cd #{app_path} && bundle exec rails generate cms:widget:video --cms_path=/website/en/_boxes"
 
           sh "cd #{app_path} && bundle exec rake cms:migrate"
         end