+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # v1.0.0
   * Renamed gem from `ice_kickstarter` to `infopark_kickstarter`.
+  * A new video box type was added. It allows to play videos from the CMS, vimeo and youtube. See
+    more details running `rails generate cms:widget:video --help`.
   * Added video obj class to the kickstart generator. This allows to upload and handle videos in a
     dedicated obj class.
   * Added rake task `rake cms:console` to open the Infopark console directly from the command line.